chore: data retrieval Added a sefe way for retrieving one's data. For now, the only information stored about a user is his username and email address.

+++ frontend/src/pages/user-page.tsx
@@ -1,15 +1,50 @@
 import { Button } from "@/components/ui/button";
 import { useToast } from "@/hooks/use-toast";
+import { useQuery } from "@tanstack/react-query";
 import axios, { AxiosError } from "axios";
+import { LoaderCircle } from "lucide-react";
 import { useNavigate } from "react-router-dom";
 
 const UserPage = () => {
   const navigate = useNavigate();
   const { toast } = useToast();
 
+  const { data, isLoading } = useQuery({
+    queryKey: ["user-info-query"],
+    queryFn: async (): Promise<{ [s: string]: string } | undefined> => {
+      try {
+        const response = await axios.get("/api/user/get-user");
+
+        return {
+          username: response.data.username,
+          email: response.data.email,
+        };
+      } catch (err) {
+        toast({
+          variant: "destructive",
+          title: "Problem with registering",
+          description: (err as AxiosError).message,
+        });
+      }
+    },
+    retryDelay: 10000,
+  });
+
+  if (isLoading) {
+    return (
+      <div className="flex items-center justify-center h-screen bg-slate-950">
+        <LoaderCircle size={38} className="animate-spin" />
+      </div>
+    );
+  }
+
   return (
-    <div className="flex items-center justify-center h-screen">
+    <div className="flex flex-col gap-4 items-center justify-center h-screen bg-slate-950">
+      <div className="flex items-center justify-center text-white">
+        {JSON.stringify(data)}
+      </div>
       <Button
+        variant="secondary"
         onClick={async () => {
           try {
             const response = await axios.delete("/api/user/logout");
